+++ CHANGELOG.md
@@ -1,6 +1,18 @@
 CHANGELOG
 ===============
 
+*V 0.2.5*
+- toast after ok click on input item dialog. 
+
+*V 0.2.4*
++ Dont select default value for spinner item if is not found on options array.
+
+*V 0.2.3*
++ Initial selected item for spinner item.
+
+*V 0.2.2*
+F selected spinner item not reflected on UI. 
+
 *V 0.2.1*
 + Input item with `withInputItem(...)`